Remove all attempts to #include pcap-int.h
We shouldn't be including it anyway as it is an "internal"
header file. Most modern pcap installs don't have it anyway
(hence we only were including it if it existed) and even those
that do (e.g. FreeBSD 10) still implement the missing pcap
functions that we would have once required it for anyway.

This may break libtrace for people using super-old versions
of the pcap library, but I would rather encourage those people
to upgrade their systems than try to make our lives difficult by
continuing to support them.