Replaced trace_get_section() with trace_get_all_metadata()
Getting the meta-data for each section individually doesn't
really make sense from a user standpoint. First, not all meta-data
records have multiple sections (ERF does, pcap-ng doesn't). Second,
future formats that support meta-data may not have any concept of
sections at all (and arguably pcap-ng doesn't even now) so the
name of the function to use may not be obvious.

So now we have trace_get_all_metadata() which is functionally
similar, but doesn't care about different sections (the section for
each meta-data field is stored with the individual item, just in
case you do need it). You call it and you get a libtrace_meta_t *
with all of the encapsulated meta-data fields from the packet in it.

This is conceptually easier to understand and requires less
format-specific knowledge on the part of the user.

Other related changes in this commit:
 * Renamed trace_get_section_option() to
   trace_get_single_meta_field(), again to make it more obvious
   what the function does.
 * Added ERF_PROV_GEN_TIME to the list of parseable ERF provenance
   fields.
 * Updated libpacketdump link_21 and link_23 modules to use
   trace_get_all_metadata() instead of trace_get_section().