traceanon: fixed issues with > 32 byte RADIUS AVPs
The digest is SHA256, so the digest buffer should not have been
increased beyond 32 bytes.

Tyler's original implementation simply zeroed any value bytes
beyond the first 32 -- I've replaced that with a continuous
repeat of the first 32 bytes. I don't think it really matters
what we do with those bytes as long as it is consistent.