Only bind ndag socket to multicast group
Previously we were binding only to the port, which meant that we would receive all UDP traffic to the bound port on the given interface. This is especially problematic when we have other ndag clients on the same machine which have joined other multicast groups as then all clients would receive traffic (to the bound port) for all groups.