rpms: be more explicit about our reliance on dpdk-wand
Apparently, yum is not smart about recursive explicit dependencies
and so will decide that it needs to install both dpdk and
dpdk-wand whenever the user asks to install libpacketdump or
libtrace-tools. This inevitably fails because the two dpdk
packages conflict.

By making all of the sub-packages depend explicitly on dpdk-wand,
hopefully this problem will go away?