Fix bad "remaining" value for corsaro-tagged packets when wire length was small
Capture length vs wire length considerations (i.e. choosing the
smaller of the two) will now only happen at layer 2 onwards, not
at the meta-data layer.

This is because meta-data is often added to a captured packet,
and therefore will extend the capture length but leave the wire
length as it was (the meta data was never "on" the wire).

Therefore, trying to correct for packets with extra capture
padding at the meta layer is incorrect, and was causing issues
where corsaro-tagged packets were appearing to have
truncated content because the value we were returning as
"remaining" had been adjusted to match the smaller wire length.