DPDK: Add vdev support and update threading for DPDK 19.11
DPDK 19.11 removed access to a number of private structures concerning
thread management. Libtrace previously modified DPDK's internal structures
and added threads pmd threads it created to these structures.

Now we track this in purely in libtrace and don't worry about
updating DPDK's structures. This is the same approach OvS takes,
so I'm fairly happy this will work. More testing still needed.

Adds support for DPDK vdevs: This allows for virtual DPDK devices.
For example net_pcap and net_tap, which are very convenient for
debugging and testing when hardware support is not available.