Add helper functions for parsing RADIUS messages
These functions are based on those written originally by
Jacob (with a few minor tweaks from me).

The RADIUS parsing does have the rather significant caveat that
it is not easy for libtrace to determine whether a packet is
actually RADIUS or not, so we're relying on the user a bit to
avoid passing in non-RADIUS packets.

There is a very basic heuristic to rule out packets where the
length in the "RADIUS header" is non-sensical, but this is not
a reliable solution (plenty of application protocols put a length
field in those bytes, for instance).