Fix pfring format code to actually compile
The ZC version is still incomplete, but at least it won't prevent
a successful build.

There's still plenty of config options that are not explicitly
handled, so we get a bunch of warnings for those still.

The non-ZC code should probably still be tested before we start
making any noise about its existence in any future releases.