XDP fix to correctly free umem buffer, Fix: #144
Patch to release the packet buffer that came from the RX queue
back to the fill queue. Previously we did not update the fill
queue. So the code was working on the assumption that packets
were always returned in the same order both by the kernel and
libtrace application.

Remaining:
Currently the fin_packet code is not thread safe, so the thread
which received the packet needs to be the one that also frees it.
While libtrace doesn't guarantee this will work for other formats,
if we can do it without a large performance hit we should try.