Fixes for trace_get_outermost_vlan/mlps
Fixes some memory alignment issues + fixes other bugs.

Reading a uint32_t (or any other non-byte size) out of a packet is generally
invalid as headers within a packet are not memory aligned. This may cause
unaligned access which is disallowed by C and the compiler might base
optimisations on this assumption, which can actually break the code.
Most reads in libtrace are fine as they are via a PACKED structure.

* Replaces uint32_t reads with uint8_t reads
* Check remaining before reading
* VLAN_NOT_FOUND and MPLS_NOT_FOUND were the wrong numbers
  - VLAN_NOT_FOUND 0xFF is actually the valid VLAN 255 -> should be 0xFFFF
  - MPLS_NOT_FOUND 0xFFFF also valid -> should be 0xFFFFFFFF
* Trace_get_outermost_vlan was incorrectly also returning the full VLAN
TCI rather than just the VLAN ID portion.