Makes changes to the way erf packets are written to a file vs a DAG card
When writing to a DAG card we do not need to set the timestamp but we need to ensure packets are correctly padded
up to the next 8 byte boundary.

When writing to a file the timestamp needs to be set and we do not need to do any padding. If a RX error has occured
write out the header with no payload.