Add support for multiple DPDK traces per program
This allows a libtrace application to open multiple DPDK interfaces by
hotplugging devices.
It is recommended to use DPDK 18.11 or newer for the best hotplug support,
but 16.11 should work.

When starting multiple DPDK traces it is recommended the end user limits the
number threads and sets a coremapping because DPDK is a polling format:
For example they can provide this in the uri:
libtrace_app coremap=[0,1],perpkt_threads=2:dpdk:0000:42:00.00 \
             coremap=[2,3],perpkt_threads=2:dpdk:0000:42:00.01

Other fixes:
* If libtrace could not request memory from the correct numa_node, warn and
  retry from any numa_node
  - Improves support for older version of DPDK (pre 18.XX) which can struggle
    to find memory.
* Add a warning when the kernel presents an invalid numa_node via /sys/
  - This upsets DPDK 17.11 and stops it mounting cards
* Enable PCI devices even when started with a vdev to allow hotplugging of PCI
  devices after launch. Whitelist the PCI host bus to stop loading PCI devices.
* Correct the invalid port value to use RTE_MAX_ETHPORTS instead of 0xff
* Correct cleanup on error to use dpdk_fin_input or dpdk_fin_output to ensure
  it frees lists allocated within the format data.
* Updates the DPDK versions in do-test-build-dpdk.sh
* Remove unused args from the whitelist_device
* Remove unused nb_ports from dpdk_format_data_t