DPDK: Pass the whitelisted device to rte_eal_init()
Depending on exactly how libtrace is linked to DPDK (particularly
when linking to the shared libraries), DPDK might not have the PCI
bus driver loaded until rte_eal_init() is called. This makes
whitelisting a device before then fail.

So instead, this change passes the device to whitelist as an argument
to rte_eal_init(), so it can be called after the PCI driver is loaded,
but, before NICs are loaded.