Add a .clang-format for #96
Starting point to give a consistent format for new code.

This is loosely based on wandio's style. Namely, using 8 space indents.

I've left LVM defaults for most settings. If anyone dislikes a default,
they can add it to the file.

Once everything is looking good we can apply the format to all of the
old code.