Fix for pausing/stopping traces under Valgrind
When debugging libtrace under valgrind trace_pstop()
and trace_ppause() sometimes hangs for an extended time,
30+ seconds.

This appears to be related to valgrind being
single-threaded and not giving other threads a chance to run.

In particular, one loop valgrind gets stuck running is
in per-packet threads such as with the ring: format.

1) The perpkt thread in pread_packet checks is_halted(),
   gets a response to halt, as the trace->state == TRACE_PAUSING
2) The pause message is not yet queued, and the perpkt thread continues
   to run
3) The perpkt thread spins waiting for the pause message, but at no
   point makes a system call etc. so valgrind can schedule that
   thread for an extended period.

This fix blocks until a message is received when in then pausing state.
This saves CPU cycles and should make the pause operation faster outside
of valgrind too.

Note: another fix is adding --fair-sched=yes to the valgrind command.