Add trace_flush_output support to ring:
Adds support for trace_flush_output() to the ring format

I had a choice between blocking until all packets are sent vs. only
ensuring they are queued for output. I opted against blocking as
this matches the DAG implementation. This also matches the definition
of flushing a file, buffers are flushed to the OS but not necessarily
to disk.