DPDK: reset port stats after starting the port
Napatech DPDK does not like resetting port stats if the port is not
yet started.

Doing this before vs. after starting the port is pretty much identical.
As it is not an atomic operation performed alongside starting the port, it
can result in a little bit of error in either case.

Additionally checks the error returned and prints an error message if
resetting port stats fails.