etsilive: fix race condition with shared decoder
Fixed etsilive_get_pdu_length() to not use a globally shared
decoder for decoding packets, as decoders are not thread-safe.
Instead, we will instantiate a fresh decoder each time this
function is called (which should only be if the packet has
been explicitly copied by the user and the cached length
values have therefore been invalidated).

Added a can_hold_packet() method to etsilive which will copy the
packet but also save the cached length values (as otherwise we
would need to decode the packet to get the capture/wire length).

If the SCB that etsilive packets are arriving into has space to
spare, then we won't even copy the packet -- the expectation would
be that the user would explicitly use make_packet_safe() if
they needed it for any length of time.

Added a new member to the libtrace_packet_t structure: fmtdata.
This can be used by the format to attach format-specific data to
a packet that can't be easily derived from packet->trace (e.g.
thread-local data for the thread that read this packet).
Only etsilive uses it right now, but could be handy for other
formats in the future?