build: use pkg-config to find ncurses
ncurses can be built in a variety of configurations, but the motivating case
for Gentoo was "split tinfo" where libtinfo is no longer included within
libncurses.

Use pkg-config to find where ncurses is installed and the required libraries
needed to link against it (which will include -ltinfo if required).

We've been shipping this patch in Gentoo for quite some time.

Signed-off-by: Sam James <sam@gentoo.org>