Replace sched_yield() with sleep in packet hashing threads
In theory, sched_yield() seems like a good idea.

In practice, it seems if you have unused CPUs then it will just
end up aggressively polling because there is no other work
to yield to. It only affects some parallel input types (ndag
is the example that led me to make this change), but results
in per packet threads that burn a constant 100% CPU even
when (or especially when) there are no packets to be read
on that input.

I've put in a short sleep instead, so the CPU usage is much
lower when there's no packets to read and therefore less
alarming to end users.