ring: make all flush operations blocking
Previously, these were non-blocking but we have no useful way
of handling an EAGAIN return -- we would simply throw an error.

Ideally, we would retry but it's not immediately obvious how
you would do that in a non-blocking way inside a typical
write_packet() workflow.

So instead I've taken the safe option of blocking every time,
at least for now.