Ignore trace_set_hasher() calls on ndag inputs
ndag is self-hashing (by the ndag sender), so trying to apply
additional hashing on top of that is not a great idea.

If we start to introduce more self-hashing inputs, we may need
to think about implementing a format property that describes
whether a hasher can be used with that format.