ndag: do not use perpkt_thread_count as a receiver count
perpkt_thread_count is set by the user, but the actual receiver
count may be capped at one (e.g. if libtrace decides to run the
format in single-threaded mode).

Note: this shouldn't really happen any more. It was a side
effect of allowing users to apply a hasher to an ndag input,
which has been disabled in the previous commit. However, it's a
good idea to have a specific receiver count anyway so we
don't get caught out by this again at some point in the future.