traceucast: remove mtu CLI option
There's no need to force outgoing messages to fit into a buffer
size that matches the link MTU -- TCP will do segmentation for us,
so we don't have to truncate large packets that exceed the MTU.

Instead, let's just use a buffer that is guaranteed to fit any
packet and save ourselves a lot of potential grief.

I've also adjusted the EAGAIN handling to use an increasing
backoff for retries, and doubled the number of allowed
attempts before giving up.