ndag: rewrite packet parsing to handle TCP properly
TCP does not guarantee that any particular receive will
contain a complete message, nor that there won't be
multiple messages arriving in a single receive call.
This means that an ndag header, ERF header or single ERF
record can span multiple buffers and we need to be able to
stitch those together as needed.

We also need to handle cases where we need to ask the socket for
more data to get a complete header / record.