pcapfile: make sure nanosecond timestamp resolution is retained
Previously, we were using trace_get_timeval() to get the
timestamp for a packet that is to be written, but obviously
this timestamp is only at microsecond resolution. Changed to
use trace_get_timespec() instead and the resulting nanosecs
are only then divided by 1000 if the output trace is going
to also use microsecond resolution.