Disable automatic fedora package builds
Fedora is a pain for us to deal with -- it's always a moving
target and the build will randomly break between versions for
stupid reasons (e.g. renaming packages we depend on).

I don't feel like we have enough fedora users to justify this
amount of effort in maintaining it, but if someone else wants
to take over that side of things then I'm happy to assist
with the handover.