Add new API method: trace_construct_packet_zc()
This is a "faster" version of trace_construct_packet() for use
in situations where you don't expect to subsequently modify the
packet in any way (i.e. just perform read/get operations).

The other benefit of this method is that you can explicitly set
the wire length for the resulting packet to a value other than
the size of the buffer that is passed in, which may be useful
in some situations.