Update ring: code to use tpacket v3
In theory, this should boost ring: performance by quite a lot.

Also added the ability for formats to specify a custom
get_layer3() function, which will allow formats that include a
direct offset to the layer 3 header (e.g. ring) to use that as a
fast path instead of walking the layer 2 headers.